+++ tests/Feature/Host/UsersTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Api\Hosts;
+namespace Tests\Feature\Host;
 
 use App\Bot;
 use Illuminate\Http\Response;
